add comment function and exception

+++ src/main/java/life/majiang/community/community/dto/QuestionDTO.java
@@ -5,14 +5,14 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
-    private Integer viewCount;
+    private Long creator;
+    private Long viewCount;
     private Integer commentCount;
     private Integer likeCount;
     private User user;
